get started button always visible in landing page
