+++ src/app/(site)/onboarding/import-linkedin/import-linkedin-page-client.tsx
@@ -2,45 +2,48 @@
 
 import * as React from "react";
 import { useRouter } from "next/navigation";
-import { Navbar } from "@/components/common/Navbar";
 import { Footer } from "@/components/common/Footer";
 import { OnboardingStepper } from "@/components/pages/onboarding/OnboardingStepper";
 import { WavyPattern } from "@/components/common/WavyPattern";
-import { TemplateSelector } from "@/components/pages/onboarding/TemplateSelector";
+import { LinkedInImport } from "@/components/pages/onboarding/LinkedInImport";
 import { Button } from "@/components/ui/button";
 import { ArrowLeft } from "lucide-react";
 
-export default function SelectTemplatePage() {
+export function ImportLinkedInPageClient() {
   const router = useRouter();
-  const [selectedTemplate, setSelectedTemplate] = React.useState<string | null>(
-    null
-  );
+  const [importedUrl, setImportedUrl] = React.useState<string | null>(null);
 
-  const handleTemplateSelect = (templateId: string) => {
-    setSelectedTemplate(templateId);
+  const handleImport = (url: string) => {
+    setImportedUrl(url);
+  };
+
+  const handleRemove = () => {
+    setImportedUrl(null);
   };
 
   const handleContinue = () => {
-    if (selectedTemplate) {
-      // TODO: Save selected template and navigate to editor
-      // router.push('/editor');
+    if (importedUrl) {
+      // Navigate to template selection (Step 3)
+      router.push("/onboarding/select-template");
     }
   };
 
   const handleBack = () => {
-    // Go back to previous step (upload-cv or import-linkedin)
-    router.back();
+    router.push("/onboarding");
+  };
+
+  const handleSkip = () => {
+    // TODO: Navigate to manual entry or skip this step
+    // router.push('/onboarding/manual-entry');
   };
 
   return (
-    <div className="min-h-screen flex flex-col bg-white">
-      <Navbar />
-
+    <>
       <main className="flex-1 flex flex-col py-8 md:py-12 px-4 relative">
-        <div className="container max-w-7xl mx-auto animate-fade-in w-full">
+        <div className="container max-w-4xl mx-auto animate-fade-in w-full">
           {/* Stepper */}
           <div className="mb-6 md:mb-8">
-            <OnboardingStepper currentStep={3} totalSteps={3} />
+            <OnboardingStepper currentStep={2} totalSteps={3} />
           </div>
 
           {/* Back Button */}
@@ -56,35 +59,41 @@ export default function SelectTemplatePage() {
           {/* Main Heading */}
           <div className="text-center mb-8 md:mb-12 px-4">
             <h1 className="text-3xl md:text-4xl lg:text-5xl font-semibold mb-4 text-gray-900">
-              Job-winning templates for you.
+              Import from LinkedIn
             </h1>
             <p className="text-base md:text-lg text-gray-600 max-w-2xl mx-auto">
-              Select a professional template that matches your style
+              We&apos;ll import your profile information and automatically fill
+              in your details
             </p>
           </div>
 
-          {/* Template Selector Component */}
+          {/* LinkedIn Import Component */}
           <div className="mb-8 px-4">
-            <TemplateSelector
-              onSelect={handleTemplateSelect}
-              selectedTemplate={selectedTemplate}
-            />
+            <LinkedInImport onImport={handleImport} onRemove={handleRemove} />
           </div>
 
           {/* Action Buttons */}
           <div className="flex flex-col sm:flex-row gap-4 justify-center items-center px-4">
             <Button
               onClick={handleContinue}
-              disabled={!selectedTemplate}
+              disabled={!importedUrl}
               className="bg-primary text-primary-foreground hover:bg-primary/90 px-8 py-6 rounded-md text-base font-medium transition-all disabled:opacity-50 disabled:cursor-not-allowed w-full sm:w-auto"
             >
               Continue
+            </Button>
+
+            <Button
+              variant="ghost"
+              onClick={handleSkip}
+              className="text-gray-600 hover:text-gray-900 w-full sm:w-auto"
+            >
+              Skip for now
             </Button>
           </div>
 
           {/* Help Text */}
           <p className="text-center text-sm text-gray-500 mt-8 px-4">
-            You can change your template later in the editor
+            Your data is secure and will only be used to create your resume
           </p>
         </div>
       </main>
@@ -95,6 +104,6 @@ export default function SelectTemplatePage() {
       </div>
 
       <Footer />
-    </div>
+    </>
   );
 }
