+++ src/app/(site)/onboarding/import-linkedin/page.tsx
@@ -1,0 +1,9 @@
+import { ImportLinkedInPageClient } from "./import-linkedin-page-client";
+
+export default function ImportLinkedInPage() {
+  return (
+    <div className="min-h-screen flex flex-col bg-white">
+      <ImportLinkedInPageClient />
+    </div>
+  );
+}
