+++ src/app/(site)/onboarding/page.tsx
@@ -1,0 +1,9 @@
+import { OnboardingPageClient } from "./onboarding-page-client";
+
+export default function OnboardingPage() {
+  return (
+    <div className="min-h-screen flex flex-col bg-white">
+      <OnboardingPageClient />
+    </div>
+  );
+}
