+++ src/app/(site)/onboarding/select-template/page.tsx
@@ -1,0 +1,9 @@
+import { SelectTemplatePageClient } from "./select-template-page-client";
+
+export default function SelectTemplatePage() {
+  return (
+    <div className="min-h-screen flex flex-col bg-white">
+      <SelectTemplatePageClient />
+    </div>
+  );
+}
