+++ src/app/(site)/onboarding/upload-cv/page.tsx
@@ -1,0 +1,9 @@
+import { UploadCVPageClient } from "./upload-cv-page-client";
+
+export default function UploadCVPage() {
+  return (
+    <div className="min-h-screen flex flex-col bg-white">
+      <UploadCVPageClient />
+    </div>
+  );
+}
