+++ src/app/(site)/page.tsx
@@ -1,0 +1,5 @@
+import { LandingPageClient } from "./landing-page-client";
+
+export default function Home() {
+  return <LandingPageClient />;
+}
